Major MCP refactor: modularized codebase, removed deprecated hallucination tool, updated README
