+++ src/crawl4ai_mcp/tools.py
@@ -1,400 +1,31 @@
-"""
-MCP server for web crawling with Crawl4AI.
+"""MCP tool implementations grouped separately for maintainability."""
 
-This server provides tools to crawl websites using Crawl4AI, automatically detecting
-the appropriate crawl method based on URL type (sitemap, txt file, or regular webpage).
-Also includes AI hallucination detection and repository parsing tools using Neo4j knowledge graphs.
-"""
-from mcp.server.fastmcp import FastMCP, Context
-from sentence_transformers import CrossEncoder
-from contextlib import asynccontextmanager
-from collections.abc import AsyncIterator
-from dataclasses import dataclass
-from typing import List, Dict, Any, Optional
-from urllib.parse import urlparse, urldefrag
-from xml.etree import ElementTree
-from dotenv import load_dotenv
-from supabase import Client
-from pathlib import Path
-from tempfile import NamedTemporaryFile
-from starlette.requests import Request
-from starlette.responses import JSONResponse
-import requests
-import asyncio
+from __future__ import annotations
+
+from typing import List, Dict, Any
 import json
 import os
-import re
+import asyncio
 import concurrent.futures
-import sys
+from urllib.parse import urlparse
 
+from .server import mcp, Context
 from crawl4ai import AsyncWebCrawler, BrowserConfig, CrawlerRunConfig, CacheMode, MemoryAdaptiveDispatcher
-
-# Add knowledge_graphs folder to path for importing knowledge graph modules
-knowledge_graphs_path = Path(__file__).resolve().parent.parent / 'knowledge_graphs'
-sys.path.append(str(knowledge_graphs_path))
-
 from utils import (
-    get_supabase_client, 
-    add_documents_to_supabase, 
-    search_documents,
+    add_documents_to_supabase,
+    add_code_examples_to_supabase,
     extract_code_blocks,
     generate_code_example_summary,
-    add_code_examples_to_supabase,
+    get_supabase_client,
+    search_documents,
+    search_code_examples,
     update_source_info,
     extract_source_summary,
-    search_code_examples,
-    LLM_MAX_CONCURRENCY
+    LLM_MAX_CONCURRENCY,
 )
 
-# Import knowledge graph modules
-from knowledge_graph_validator import KnowledgeGraphValidator
-from parse_repo_into_neo4j import DirectNeo4jExtractor
-from ai_script_analyzer import AIScriptAnalyzer
-from hallucination_reporter import HallucinationReporter
+from .server import smart_chunk_markdown, extract_section_info
 
-# Load environment variables from the project root .env file
-project_root = Path(__file__).resolve().parent.parent
-dotenv_path = project_root / '.env'
-
-# Force override of existing environment variables
-load_dotenv(dotenv_path, override=True)
-
-# Helper functions for Neo4j validation and error handling
-def validate_neo4j_connection() -> bool:
-    """Check if Neo4j environment variables are configured."""
-    return all([
-        os.getenv("NEO4J_URI"),
-        os.getenv("NEO4J_USER"),
-        os.getenv("NEO4J_PASSWORD")
-    ])
-
-def format_neo4j_error(error: Exception) -> str:
-    """Format Neo4j connection errors for user-friendly messages."""
-    error_str = str(error).lower()
-    if "authentication" in error_str or "unauthorized" in error_str:
-        return "Neo4j authentication failed. Check NEO4J_USER and NEO4J_PASSWORD."
-    elif "connection" in error_str or "refused" in error_str or "timeout" in error_str:
-        return "Cannot connect to Neo4j. Check NEO4J_URI and ensure Neo4j is running."
-    elif "database" in error_str:
-        return "Neo4j database error. Check if the database exists and is accessible."
-    else:
-        return f"Neo4j error: {str(error)}"
-
-def validate_script_path(script_path: str) -> Dict[str, Any]:
-    """Validate script path and return error info if invalid."""
-    if not script_path or not isinstance(script_path, str):
-        return {"valid": False, "error": "Script path is required"}
-    
-    if not os.path.exists(script_path):
-        return {"valid": False, "error": f"Script not found: {script_path}"}
-    
-    if not script_path.endswith('.py'):
-        return {"valid": False, "error": "Only Python (.py) files are supported"}
-    
-    try:
-        # Check if file is readable
-        with open(script_path, 'r', encoding='utf-8') as f:
-            f.read(1)  # Read first character to test
-        return {"valid": True}
-    except Exception as e:
-        return {"valid": False, "error": f"Cannot read script file: {str(e)}"}
-
-def validate_github_url(repo_url: str) -> Dict[str, Any]:
-    """Validate GitHub repository URL."""
-    if not repo_url or not isinstance(repo_url, str):
-        return {"valid": False, "error": "Repository URL is required"}
-    
-    repo_url = repo_url.strip()
-    
-    # Basic GitHub URL validation
-    if not ("github.com" in repo_url.lower() or repo_url.endswith(".git")):
-        return {"valid": False, "error": "Please provide a valid GitHub repository URL"}
-    
-    # Check URL format
-    if not (repo_url.startswith("https://") or repo_url.startswith("git@")):
-        return {"valid": False, "error": "Repository URL must start with https:// or git@"}
-    
-    return {"valid": True, "repo_name": repo_url.split('/')[-1].replace('.git', '')}
-
-# Create a dataclass for our application context
-@dataclass
-class Crawl4AIContext:
-    """Context for the Crawl4AI MCP server."""
-    crawler: AsyncWebCrawler
-    supabase_client: Client
-    reranking_model: Optional[CrossEncoder] = None
-    knowledge_validator: Optional[Any] = None  # KnowledgeGraphValidator when available
-    repo_extractor: Optional[Any] = None       # DirectNeo4jExtractor when available
-
-@asynccontextmanager
-async def crawl4ai_lifespan(server: FastMCP) -> AsyncIterator[Crawl4AIContext]:
-    """
-    Manages the Crawl4AI client lifecycle.
-    
-    Args:
-        server: The FastMCP server instance
-        
-    Yields:
-        Crawl4AIContext: The context containing the Crawl4AI crawler and Supabase client
-    """
-    # Create browser configuration
-    browser_config = BrowserConfig(
-        headless=True,
-        verbose=False
-    )
-    
-    # Initialize the crawler
-    crawler = AsyncWebCrawler(config=browser_config)
-    await crawler.__aenter__()
-    
-    # Initialize Supabase client
-    supabase_client = get_supabase_client()
-    
-    # Initialize cross-encoder model for reranking if enabled
-    reranking_model = None
-    if os.getenv("USE_RERANKING", "false") == "true":
-        try:
-            reranking_model = CrossEncoder("cross-encoder/ms-marco-MiniLM-L-6-v2")
-        except Exception as e:
-            print(f"Failed to load reranking model: {e}")
-            reranking_model = None
-    
-    # Initialize Neo4j components if configured and enabled
-    knowledge_validator = None
-    repo_extractor = None
-    
-    # Check if knowledge graph functionality is enabled
-    knowledge_graph_enabled = os.getenv("USE_KNOWLEDGE_GRAPH", "false") == "true"
-    
-    if knowledge_graph_enabled:
-        neo4j_uri = os.getenv("NEO4J_URI")
-        neo4j_user = os.getenv("NEO4J_USER")
-        neo4j_password = os.getenv("NEO4J_PASSWORD")
-        
-        if neo4j_uri and neo4j_user and neo4j_password:
-            try:
-                print("Initializing knowledge graph components...")
-                
-                # Initialize knowledge graph validator
-                knowledge_validator = KnowledgeGraphValidator(neo4j_uri, neo4j_user, neo4j_password)
-                await knowledge_validator.initialize()
-                print("✓ Knowledge graph validator initialized")
-                
-                # Initialize repository extractor
-                repo_extractor = DirectNeo4jExtractor(neo4j_uri, neo4j_user, neo4j_password)
-                await repo_extractor.initialize()
-                print("✓ Repository extractor initialized")
-                
-            except Exception as e:
-                print(f"Failed to initialize Neo4j components: {format_neo4j_error(e)}")
-                knowledge_validator = None
-                repo_extractor = None
-        else:
-            print("Neo4j credentials not configured - knowledge graph tools will be unavailable")
-    else:
-        print("Knowledge graph functionality disabled - set USE_KNOWLEDGE_GRAPH=true to enable")
-    
-    context = Crawl4AIContext(
-        crawler=crawler,
-        supabase_client=supabase_client,
-        reranking_model=reranking_model,
-        knowledge_validator=knowledge_validator,
-        repo_extractor=repo_extractor,
-    )
-    # expose the lifespan context on the FastMCP instance so HTTP routes can
-    # access it without an active MCP request context
-    server.lifespan_context = context
-
-    try:
-        yield context
-    finally:
-        # remove reference for cleanup
-        server.lifespan_context = None
-        # Clean up all components
-        await crawler.__aexit__(None, None, None)
-        if knowledge_validator:
-            try:
-                await knowledge_validator.close()
-                print("✓ Knowledge graph validator closed")
-            except Exception as e:
-                print(f"Error closing knowledge validator: {e}")
-        if repo_extractor:
-            try:
-                await repo_extractor.close()
-                print("✓ Repository extractor closed")
-            except Exception as e:
-                print(f"Error closing repository extractor: {e}")
-
-# Initialize FastMCP server
-mcp = FastMCP(
-    "mcp-crawl4ai-rag",
-    description="MCP server for RAG and web crawling with Crawl4AI",
-    lifespan=crawl4ai_lifespan,
-    host=os.getenv("HOST", "0.0.0.0"),
-    port=os.getenv("PORT", "8051")
-)
-
-def rerank_results(model: CrossEncoder, query: str, results: List[Dict[str, Any]], content_key: str = "content") -> List[Dict[str, Any]]:
-    """
-    Rerank search results using a cross-encoder model.
-    
-    Args:
-        model: The cross-encoder model to use for reranking
-        query: The search query
-        results: List of search results
-        content_key: The key in each result dict that contains the text content
-        
-    Returns:
-        Reranked list of results
-    """
-    if not model or not results:
-        return results
-    
-    try:
-        # Extract content from results
-        texts = [result.get(content_key, "") for result in results]
-        
-        # Create pairs of [query, document] for the cross-encoder
-        pairs = [[query, text] for text in texts]
-        
-        # Get relevance scores from the cross-encoder
-        scores = model.predict(pairs)
-        
-        # Add scores to results and sort by score (descending)
-        for i, result in enumerate(results):
-            result["rerank_score"] = float(scores[i])
-        
-        # Sort by rerank score
-        reranked = sorted(results, key=lambda x: x.get("rerank_score", 0), reverse=True)
-        
-        return reranked
-    except Exception as e:
-        print(f"Error during reranking: {e}")
-        return results
-
-def is_sitemap(url: str) -> bool:
-    """
-    Check if a URL is a sitemap.
-    
-    Args:
-        url: URL to check
-        
-    Returns:
-        True if the URL is a sitemap, False otherwise
-    """
-    return url.endswith('sitemap.xml') or 'sitemap' in urlparse(url).path
-
-def is_txt(url: str) -> bool:
-    """
-    Check if a URL is a text file.
-    
-    Args:
-        url: URL to check
-        
-    Returns:
-        True if the URL is a text file, False otherwise
-    """
-    return url.endswith('.txt')
-
-def parse_sitemap(sitemap_url: str) -> List[str]:
-    """
-    Parse a sitemap and extract URLs.
-    
-    Args:
-        sitemap_url: URL of the sitemap
-        
-    Returns:
-        List of URLs found in the sitemap
-    """
-    resp = requests.get(sitemap_url)
-    urls = []
-
-    if resp.status_code == 200:
-        try:
-            tree = ElementTree.fromstring(resp.content)
-            urls = [loc.text for loc in tree.findall('.//{*}loc')]
-        except Exception as e:
-            print(f"Error parsing sitemap XML: {e}")
-
-    return urls
-
-def smart_chunk_markdown(text: str, chunk_size: int = 5000) -> List[str]:
-    """Split text into chunks, respecting code blocks and paragraphs."""
-    chunks = []
-    start = 0
-    text_length = len(text)
-
-    while start < text_length:
-        # Calculate end position
-        end = start + chunk_size
-
-        # If we're at the end of the text, just take what's left
-        if end >= text_length:
-            chunks.append(text[start:].strip())
-            break
-
-        # Try to find a code block boundary first (```)
-        chunk = text[start:end]
-        code_block = chunk.rfind('```')
-        if code_block != -1 and code_block > chunk_size * 0.3:
-            end = start + code_block
-
-        # If no code block, try to break at a paragraph
-        elif '\n\n' in chunk:
-            # Find the last paragraph break
-            last_break = chunk.rfind('\n\n')
-            if last_break > chunk_size * 0.3:  # Only break if we're past 30% of chunk_size
-                end = start + last_break
-
-        # If no paragraph break, try to break at a sentence
-        elif '. ' in chunk:
-            # Find the last sentence break
-            last_period = chunk.rfind('. ')
-            if last_period > chunk_size * 0.3:  # Only break if we're past 30% of chunk_size
-                end = start + last_period + 1
-
-        # Extract chunk and clean it up
-        chunk = text[start:end].strip()
-        if chunk:
-            chunks.append(chunk)
-
-        # Move start position for next chunk
-        start = end
-
-    return chunks
-
-def extract_section_info(chunk: str) -> Dict[str, Any]:
-    """
-    Extracts headers and stats from a chunk.
-    
-    Args:
-        chunk: Markdown chunk
-        
-    Returns:
-        Dictionary with headers and stats
-    """
-    headers = re.findall(r'^(#+)\s+(.+)$', chunk, re.MULTILINE)
-    header_str = '; '.join([f'{h[0]} {h[1]}' for h in headers]) if headers else ''
-
-    return {
-        "headers": header_str,
-        "char_count": len(chunk),
-        "word_count": len(chunk.split())
-    }
-
-def process_code_example(args):
-    """
-    Process a single code example to generate its summary.
-    This function is designed to be used with concurrent.futures.
-    
-    Args:
-        args: Tuple containing (code, context_before, context_after)
-        
-    Returns:
-        The generated summary
-    """
-    code, context_before, context_after = args
-    return generate_code_example_summary(code, context_before, context_after)
 
 @mcp.tool()
 async def crawl_single_page(ctx: Context, url: str) -> str:
@@ -533,9 +164,6 @@ async def crawl_single_page(ctx: Context, url: str) -> str:
         return json.dumps({
             "success": False,
             "url": url,
-            "error": str(e)
-        }, indent=2)
-
 @mcp.tool()
 async def smart_crawl_url(ctx: Context, url: str, max_depth: int = 3, max_concurrent: int = 10, chunk_size: int = 5000) -> str:
     """
@@ -726,12 +354,6 @@ async def smart_crawl_url(ctx: Context, url: str, max_depth: int = 3, max_concur
             "urls_crawled": [doc['url'] for doc in crawl_results][:5] + (["..."] if len(crawl_results) > 5 else [])
         }, indent=2)
     except Exception as e:
-        return json.dumps({
-            "success": False,
-            "url": url,
-            "error": str(e)
-        }, indent=2)
-
 @mcp.tool()
 async def get_available_sources(ctx: Context) -> str:
     """
@@ -780,10 +402,6 @@ async def get_available_sources(ctx: Context) -> str:
     except Exception as e:
         return json.dumps({
             "success": False,
-            "error": str(e)
-        }, indent=2)
-
-@mcp.tool()
 async def perform_rag_query(ctx: Context, query: str, source: str = None, match_count: int = 5) -> str:
     """
     Perform a RAG (Retrieval Augmented Generation) query on the stored content.
@@ -922,7 +540,6 @@ async def perform_rag_query(ctx: Context, query: str, source: str = None, match_
             "error": str(e)
         }, indent=2)
 
-@mcp.tool()
 async def search_code_examples(ctx: Context, query: str, source_id: str = None, match_count: int = 5) -> str:
     """
     Search for code examples relevant to the query.
@@ -1080,314 +697,6 @@ async def search_code_examples(ctx: Context, query: str, source_id: str = None, 
         }, indent=2)
 
 @mcp.tool()
-async def check_ai_script_hallucinations(
-    ctx: Context,
-    script_path: str | None = None,
-    script_content: str | None = None,
-    filename: str | None = None,
-) -> str:
-    """
-    Check an AI-generated Python script for hallucinations using the knowledge graph.
-    
-    This tool analyzes a Python script for potential AI hallucinations by validating
-    imports, method calls, class instantiations, and function calls against a Neo4j
-    knowledge graph containing real repository data.
-    
-    The tool performs comprehensive analysis including:
-    - Import validation against known repositories
-    - Method call validation on classes from the knowledge graph
-    - Class instantiation parameter validation
-    - Function call parameter validation
-    - Attribute access validation
-    
-    Args:
-        ctx: The MCP server provided context
-        script_path: Absolute path to the Python script to analyze. Optional if
-            ``script_content`` is provided.
-        script_content: Raw Python source to validate. If provided, it will be
-            written to a temporary file and analyzed as ``filename``.
-        filename: Optional filename to use when creating the temporary file.
-    
-    Returns:
-        JSON string with hallucination detection results, confidence scores, and recommendations
-    """
-    try:
-        # Check if knowledge graph functionality is enabled
-        knowledge_graph_enabled = os.getenv("USE_KNOWLEDGE_GRAPH", "false") == "true"
-        if not knowledge_graph_enabled:
-            return json.dumps({
-                "success": False,
-                "error": "Knowledge graph functionality is disabled. Set USE_KNOWLEDGE_GRAPH=true in environment."
-            }, indent=2)
-        
-        # Get the knowledge validator from context. When invoked via MCP tools
-        # there will be an active request_context. HTTP routes call this
-        # function without one, so fall back to the server's lifespan_context.
-        knowledge_validator = None
-        if ctx._request_context is not None:
-            knowledge_validator = ctx.request_context.lifespan_context.knowledge_validator
-        elif hasattr(ctx.fastmcp, "lifespan_context"):
-            knowledge_validator = getattr(ctx.fastmcp.lifespan_context, "knowledge_validator", None)
-        
-        if not knowledge_validator:
-            return json.dumps({
-                "success": False,
-                "error": "Knowledge graph validator not available. Check Neo4j configuration in environment variables."
-            }, indent=2)
-        
-        tmp_path = None
-
-        if script_content is not None:
-            suffix = filename if (filename and filename.endswith(".py")) else (
-                (filename or "temp_script") + ".py")
-            with NamedTemporaryFile("w", suffix=suffix, delete=False, dir="/tmp") as tmp:
-                tmp.write(script_content)
-                tmp_path = tmp.name
-            path_to_use = tmp_path
-        else:
-            path_to_use = script_path
-
-        # Validate script path
-        validation = validate_script_path(path_to_use)
-        if not validation["valid"]:
-            if tmp_path:
-                os.remove(tmp_path)
-            return json.dumps({
-                "success": False,
-                "script_path": path_to_use,
-                "error": validation["error"]
-            }, indent=2)
-
-        # Step 1: Analyze script structure using AST
-        analyzer = AIScriptAnalyzer()
-        analysis_result = analyzer.analyze_script(path_to_use)
-        
-        if analysis_result.errors:
-            print(f"Analysis warnings for {path_to_use}: {analysis_result.errors}")
-        
-        # Step 2: Validate against knowledge graph
-        validation_result = await knowledge_validator.validate_script(analysis_result)
-        
-        # Step 3: Generate comprehensive report
-        reporter = HallucinationReporter()
-        report = reporter.generate_comprehensive_report(validation_result)
-        
-        # Format response with comprehensive information
-        result_json = json.dumps({
-            "success": True,
-            "script_path": path_to_use,
-            "overall_confidence": validation_result.overall_confidence,
-            "validation_summary": {
-                "total_validations": report["validation_summary"]["total_validations"],
-                "valid_count": report["validation_summary"]["valid_count"],
-                "invalid_count": report["validation_summary"]["invalid_count"],
-                "uncertain_count": report["validation_summary"]["uncertain_count"],
-                "not_found_count": report["validation_summary"]["not_found_count"],
-                "hallucination_rate": report["validation_summary"]["hallucination_rate"]
-            },
-            "hallucinations_detected": report["hallucinations_detected"],
-            "recommendations": report["recommendations"],
-            "analysis_metadata": {
-                "total_imports": report["analysis_metadata"]["total_imports"],
-                "total_classes": report["analysis_metadata"]["total_classes"],
-                "total_methods": report["analysis_metadata"]["total_methods"],
-                "total_attributes": report["analysis_metadata"]["total_attributes"],
-                "total_functions": report["analysis_metadata"]["total_functions"]
-            },
-            "libraries_analyzed": report.get("libraries_analyzed", [])
-        }, indent=2)
-        if tmp_path:
-            os.remove(tmp_path)
-        return result_json
-        
-    except Exception as e:
-        if tmp_path:
-            os.remove(tmp_path)
-        return json.dumps({
-            "success": False,
-            "script_path": path_to_use if 'path_to_use' in locals() else str(script_path),
-            "error": f"Analysis failed: {str(e)}"
-        }, indent=2)
-
-
-@mcp.custom_route("/api/check_script_hallucinations", methods=["POST"])
-async def check_script_hallucinations_api(request: Request):
-    """HTTP endpoint to validate Python code content for hallucinations."""
-    try:
-        payload = await request.json()
-    except Exception:
-        return JSONResponse({"success": False, "error": "Invalid JSON body"}, status_code=400)
-
-    script_content = payload.get("script_content")
-    filename = payload.get("filename", "temp_script.py")
-    if not script_content:
-        return JSONResponse({"success": False, "error": "script_content is required"}, status_code=400)
-
-    # No MCP request context is active for HTTP routes, so create a Context
-    # with only the FastMCP reference. check_ai_script_hallucinations will
-    # fall back to the server's lifespan_context in this case.
-    ctx = Context(fastmcp=mcp)
-    result_json = await check_ai_script_hallucinations(
-        ctx,
-        script_content=script_content,
-        filename=filename,
-    )
-    return JSONResponse(json.loads(result_json))
-
-@mcp.tool()
-async def query_knowledge_graph(ctx: Context, command: str) -> str:
-    """
-    Query and explore the Neo4j knowledge graph containing repository data.
-    
-    This tool provides comprehensive access to the knowledge graph for exploring repositories,
-    classes, methods, functions, and their relationships. Perfect for understanding what data
-    is available for hallucination detection and debugging validation results.
-    
-    **⚠️ IMPORTANT: Always start with the `repos` command first!**
-    Before using any other commands, run `repos` to see what repositories are available
-    in your knowledge graph. This will help you understand what data you can explore.
-    
-    ## Available Commands:
-    
-    **Repository Commands:**
-    - `repos` - **START HERE!** List all repositories in the knowledge graph
-    - `explore <repo_name>` - Get detailed overview of a specific repository
-    
-    **Class Commands:**  
-    - `classes` - List all classes across all repositories (limited to 20)
-    - `classes <repo_name>` - List classes in a specific repository
-    - `class <class_name>` - Get detailed information about a specific class including methods and attributes
-    
-    **Method Commands:**
-    - `method <method_name>` - Search for methods by name across all classes
-    - `method <method_name> <class_name>` - Search for a method within a specific class
-
-    **Function Commands:**
-    - `function <function_name>` - Search for standalone functions by name
-    
-    **Custom Query:**
-    - `query <cypher_query>` - Execute a custom Cypher query (results limited to 20 records)
-    
-    ## Knowledge Graph Schema:
-    
-    **Node Types:**
-    - Repository: `(r:Repository {name: string})`
-    - File: `(f:File {path: string, module_name: string})`
-    - Class: `(c:Class {name: string, full_name: string})`
-    - Method: `(m:Method {name: string, params_list: [string], params_detailed: [string], return_type: string, args: [string]})`
-    - Function: `(func:Function {name: string, params_list: [string], params_detailed: [string], return_type: string, args: [string]})`
-    - Attribute: `(a:Attribute {name: string, type: string})`
-    
-    **Relationships:**
-    - `(r:Repository)-[:CONTAINS]->(f:File)`
-    - `(f:File)-[:DEFINES]->(c:Class)`
-    - `(c:Class)-[:HAS_METHOD]->(m:Method)`
-    - `(c:Class)-[:HAS_ATTRIBUTE]->(a:Attribute)`
-    - `(f:File)-[:DEFINES]->(func:Function)`
-    
-    ## Example Workflow:
-    ```
-    1. repos                                    # See what repositories are available
-    2. explore pydantic-ai                      # Explore a specific repository
-    3. classes pydantic-ai                      # List classes in that repository
-    4. class Agent                              # Explore the Agent class
-    5. method run_stream                        # Search for run_stream method
-    6. method __init__ Agent                    # Find Agent constructor
-    7. query "MATCH (c:Class)-[:HAS_METHOD]->(m:Method) WHERE m.name = 'run' RETURN c.name, m.name LIMIT 5"
-    ```
-    
-    Args:
-        ctx: The MCP server provided context
-        command: Command string to execute (see available commands above)
-    
-    Returns:
-        JSON string with query results, statistics, and metadata
-    """
-    try:
-        # Check if knowledge graph functionality is enabled
-        knowledge_graph_enabled = os.getenv("USE_KNOWLEDGE_GRAPH", "false") == "true"
-        if not knowledge_graph_enabled:
-            return json.dumps({
-                "success": False,
-                "error": "Knowledge graph functionality is disabled. Set USE_KNOWLEDGE_GRAPH=true in environment."
-            }, indent=2)
-        
-        # Get Neo4j driver from context
-        repo_extractor = ctx.request_context.lifespan_context.repo_extractor
-        if not repo_extractor or not repo_extractor.driver:
-            return json.dumps({
-                "success": False,
-                "error": "Neo4j connection not available. Check Neo4j configuration in environment variables."
-            }, indent=2)
-        
-        # Parse command
-        command = command.strip()
-        if not command:
-            return json.dumps({
-                "success": False,
-                "command": "",
-                "error": "Command cannot be empty. Available commands: repos, explore <repo>, classes [repo], class <name>, method <name> [class], function <name>, query <cypher>"
-            }, indent=2)
-        
-        parts = command.split()
-        cmd = parts[0].lower()
-        args = parts[1:] if len(parts) > 1 else []
-        
-        async with repo_extractor.driver.session() as session:
-            # Route to appropriate handler
-            if cmd == "repos":
-                return await _handle_repos_command(session, command)
-            elif cmd == "explore":
-                if not args:
-                    return json.dumps({
-                        "success": False,
-                        "command": command,
-                        "error": "Repository name required. Usage: explore <repo_name>"
-                    }, indent=2)
-                return await _handle_explore_command(session, command, args[0])
-            elif cmd == "classes":
-                repo_name = args[0] if args else None
-                return await _handle_classes_command(session, command, repo_name)
-            elif cmd == "class":
-                if not args:
-                    return json.dumps({
-                        "success": False,
-                        "command": command,
-                        "error": "Class name required. Usage: class <class_name>"
-                    }, indent=2)
-                return await _handle_class_command(session, command, args[0])
-            elif cmd == "method":
-                if not args:
-                    return json.dumps({
-                        "success": False,
-                        "command": command,
-                        "error": "Method name required. Usage: method <method_name> [class_name]"
-                    }, indent=2)
-                method_name = args[0]
-                class_name = args[1] if len(args) > 1 else None
-                return await _handle_method_command(session, command, method_name, class_name)
-            elif cmd == "function":
-                if not args:
-                    return json.dumps({
-                        "success": False,
-                        "command": command,
-                        "error": "Function name required. Usage: function <function_name>"
-                    }, indent=2)
-                function_name = args[0]
-                return await _handle_function_command(session, command, function_name)
-            elif cmd == "query":
-                if not args:
-                    return json.dumps({
-                        "success": False,
-                        "command": command,
-                        "error": "Cypher query required. Usage: query <cypher_query>"
-                    }, indent=2)
-                cypher_query = " ".join(args)
-                return await _handle_query_command(session, command, cypher_query)
-            else:
-                return json.dumps({
-                    "success": False,
-                    "command": command,
                     "error": f"Unknown command '{cmd}'. Available commands: repos, explore <repo>, classes [repo], class <name>, method <name> [class], function <name>, query <cypher>"
                 }, indent=2)
                 
@@ -1861,12 +1170,10 @@ async def parse_github_repository(ctx: Context, repo_url: str) -> str:
             "statistics": stats,
             "ready_for_validation": True,
             "next_steps": [
-                "Repository is now available for hallucination detection",
-                f"Use check_ai_script_hallucinations to validate scripts against {repo_name}",
+                "Repository is now available for validation against the knowledge graph",
                 "The knowledge graph contains classes, methods, and functions from this repository"
             ]
         }, indent=2)
-        
     except Exception as e:
         return json.dumps({
             "success": False,
